+++ src/com/dakoda/alr/game/TextRPG.java
@@ -4,6 +4,8 @@ import com.dakoda.alr.game.quest.Quest;
 public class TextRPG {
 
     public static void main(String[] args) {
-        Quest quest = new Quest.QuestBuilder().build();
+        Quest quest = new Quest.QuestBuilder()
+                .addItemRequirement(null, 1)
+                .build();
     }
 }
